+++ process/core/io/plot/summary.py
@@ -58,7 +58,6 @@ from process.models.physics.current_drive import (
 from process.models.physics.impurity_radiation import read_impurity_file
 from process.models.physics.l_h_transition import PlasmaConfinementTransitionModel
 from process.models.physics.physics import BetaComponentLimits, BetaNormMaxModel
-from process.models.physics.physics import DetailedPhysics
 from process.models.physics.plasma_current import (
     PlasmaCurrentModel,
     PlasmaDiamagneticCurrentModel,
@@ -12820,6 +12819,737 @@ def plot_ebw_ecrh_coupling_graph(axis: plt.Axes, mfile: MFile, scan: int):
     axis.minorticks_on()
 
 
+def plot_larmor_radius_profile(axis: plt.Axes, mfile_data: MFile, scan: int):
+    """Plot the Larmor radius profile on the given axis."""
+    radius_plasma_deuteron_larmor_profile = [
+        mfile_data.data[
+            f"radius_plasma_deuteron_toroidal_larmor_isotropic_profile{i}"
+        ].get_scan(scan)
+        for i in range(
+            2 * int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan))
+        )
+    ]
+
+    radius_plasma_triton_larmor_profile = [
+        mfile_data.data[
+            f"radius_plasma_triton_toroidal_larmor_isotropic_profile{i}"
+        ].get_scan(scan)
+        for i in range(
+            2 * int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan))
+        )
+    ]
+
+    radius_plasma_deuteron_larmor_profile_mm = [
+        radius * 1e3 for radius in radius_plasma_deuteron_larmor_profile
+    ]
+
+    radius_plasma_triton_larmor_profile_mm = [
+        radius * 1e3 for radius in radius_plasma_triton_larmor_profile
+    ]
+
+    axis.plot(
+        np.linspace(-1, 1, len(radius_plasma_deuteron_larmor_profile_mm)),
+        radius_plasma_deuteron_larmor_profile_mm,
+        color="red",
+        linestyle="-",
+        label=r"$\rho_{Larmor,toroidal,D}$",
+    )
+
+    axis.plot(
+        np.linspace(-1, 1, len(radius_plasma_triton_larmor_profile_mm)),
+        radius_plasma_triton_larmor_profile_mm,
+        color="green",
+        linestyle="-",
+        label=r"$\rho_{Larmor,toroidal,T}$",
+    )
+
+    axis.set_ylabel(r"Larmor Radii [mm]")
+    axis.set_title(r" Toroidal Larmor Radii ($v_{\perp}^2 = 2v_{th}^2$)")
+    axis.set_xlabel("$\\rho \\ [r/a]$")
+    axis.grid(True, which="both", linestyle="--", alpha=0.5)
+    axis.minorticks_on()
+    axis.legend()
+
+
+def plot_debye_length_profile(axis: plt.Axes, mfile_data: MFile, scan: int):
+    """Plot the Debye length profile on the given axis.
+
+    Parameters
+    ----------
+    axis : plt.Axes
+        Axis object to plot on
+    mfile_data : mf.MFile
+        MFILE data object
+    scan : int
+        Scan number to use
+
+    """
+    len_plasma_debye_electron_profile = [
+        mfile_data.data[f"len_plasma_debye_electron_profile{i}"].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    # Convert to micrometres (1e-6 m)
+    len_plasma_debye_electron_profile_um = [
+        length * 1e6 for length in len_plasma_debye_electron_profile
+    ]
+
+    axis.plot(
+        np.linspace(0, 1, len(len_plasma_debye_electron_profile_um)),
+        len_plasma_debye_electron_profile_um,
+        color="blue",
+        linestyle="-",
+        label=r"$\lambda_{Debye,e}$",
+    )
+
+    axis.set_ylabel(r"Debye Length [$\mu$m]")
+
+    axis.set_xlabel("$\\rho \\ [r/a]$")
+    axis.grid(True, which="both", linestyle="--", alpha=0.5)
+    axis.set_xlim([0, 1.025])
+    axis.minorticks_on()
+    axis.legend()
+
+
+def plot_velocity_profile(axis: plt.Axes, mfile_data: MFile, scan: int) -> None:
+    """Plot the electron thermal velocity profile on the given axis.
+
+    Parameters
+    ----------
+    axis : plt.Axes
+        Axis object to plot on
+    mfile_data : mf.MFile
+        MFILE data object
+    scan : int
+        Scan number to use
+    """
+
+    vel_plasma_electron_profile = [
+        mfile_data.data[f"vel_plasma_electron_profile{i}"].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+    vel_plasma_deuteron_profile = [
+        mfile_data.data[f"vel_plasma_deuteron_profile{i}"].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+    vel_plasma_triton_profile = [
+        mfile_data.data[f"vel_plasma_triton_profile{i}"].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+    vel_plasma_alpha_thermal_profile = [
+        mfile_data.data[f"vel_plasma_alpha_thermal_profile{i}"].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    vel_plasma_alpha_birth = mfile_data.data["vel_plasma_alpha_birth"].get_scan(scan)
+
+    axis.plot(
+        np.linspace(0, 1, len(vel_plasma_electron_profile)),
+        vel_plasma_electron_profile,
+        color="blue",
+        linestyle="-",
+        label=r"$v_{e}$",
+    )
+    axis.plot(
+        np.linspace(0, 1, len(vel_plasma_deuteron_profile)),
+        vel_plasma_deuteron_profile,
+        color="pink",
+        linestyle="-",
+        label=r"$v_{D}$",
+    )
+    axis.plot(
+        np.linspace(0, 1, len(vel_plasma_triton_profile)),
+        vel_plasma_triton_profile,
+        color="green",
+        linestyle="-",
+        label=r"$v_{T}$",
+    )
+    axis.plot(
+        np.linspace(0, 1, len(vel_plasma_alpha_thermal_profile)),
+        vel_plasma_alpha_thermal_profile,
+        color="red",
+        linestyle="-",
+        label=r"$v_{\alpha,thermal}$",
+    )
+    axis.axhline(
+        vel_plasma_alpha_birth,
+        color="red",
+        linestyle="--",
+        linewidth=1.5,
+        label=r"$v_{\alpha,birth}$",
+    )
+
+    axis.set_yscale("log")
+    axis.set_ylabel("Velocity [m/s]")
+    axis.set_xlabel("$\\rho \\ [r/a]$")
+    axis.grid(True, which="both", linestyle="--", alpha=0.5)
+    axis.set_xlim([0, 1.025])
+    axis.minorticks_on()
+    axis.legend()
+
+
+def plot_electron_frequency_profile(
+    axis: plt.Axes, mfile_data: MFile, scan: int
+) -> None:
+    """Plot the electron thermal frequency profile on the given axis.
+
+    Parameters
+    ----------
+    axis : plt.Axes
+        Axis object to plot on
+    mfile_data : MFile
+        MFILE data object
+    scan : int
+        Scan number to use
+    """
+
+    freq_plasma_electron_profile = [
+        mfile_data.data[f"freq_plasma_electron_profile{i}"].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+    freq_plasma_larmor_toroidal_electron_profile = [
+        mfile_data.data[f"freq_plasma_larmor_toroidal_electron_profile{i}"].get_scan(
+            scan
+        )
+        for i in range(
+            2 * int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan))
+        )
+    ]
+
+    freq_plasma_upper_hybrid_electron_profile = [
+        mfile_data.data[f"freq_plasma_upper_hybrid_profile{i}"].get_scan(scan)
+        for i in range(
+            2 * int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan))
+        )
+    ]
+
+    axis.plot(
+        np.linspace(-1, 1, len(freq_plasma_larmor_toroidal_electron_profile)),
+        np.array(freq_plasma_larmor_toroidal_electron_profile) / 1e9,
+        color="red",
+        linestyle="-",
+        label=r"$f_{Larmor,toroidal,e}$ | Fundamental",
+    )
+
+    axis.plot(
+        np.linspace(-1, 1, len(freq_plasma_larmor_toroidal_electron_profile)),
+        2 * np.array(freq_plasma_larmor_toroidal_electron_profile) / 1e9,
+        color="red",
+        linestyle="--",
+        label=r"$f_{Larmor,toroidal,e}$ | 2nd harmonic",
+    )
+
+    axis.plot(
+        np.linspace(-1, 1, len(freq_plasma_larmor_toroidal_electron_profile)),
+        3 * np.array(freq_plasma_larmor_toroidal_electron_profile) / 1e9,
+        color="red",
+        linestyle=":",
+        label=r"$f_{Larmor,toroidal,e}$ | 3rd harmonic",
+    )
+
+    x = np.linspace(0, 1, len(freq_plasma_electron_profile))
+    y = np.array(freq_plasma_electron_profile) / 1e9
+    # original curve
+    axis.plot(
+        x, y, color="blue", linestyle="-", label=r"$\omega_{p,e}$ | Plasma Frequency"
+    )
+    # mirrored across the y-axis (drawn at negative rho)
+    axis.plot(-x, y, color="blue", linestyle="-", label="_nolegend_")
+
+    axis.plot(
+        np.linspace(-1, 1, len(freq_plasma_upper_hybrid_electron_profile)),
+        np.array(freq_plasma_upper_hybrid_electron_profile) / 1e9,
+        color="purple",
+        linestyle="-",
+        label=r"$\omega_{UH,e}$ | Upper Hybrid",
+    )
+
+    axis.set_xlim(-1.025, 1.025)
+    axis.set_ylim(None, max(freq_plasma_larmor_toroidal_electron_profile) / 1e9 * 1.6)
+
+    axis.set_xlabel("$\\rho$ [r/a]")
+    axis.set_ylabel("Frequency [GHz]")
+    axis.grid(True, which="both", linestyle="--", alpha=0.5)
+
+    # Add secondary x-axis showing radius in metres below the primary axis
+    ax2 = axis.twiny()
+    rmajor = mfile_data.get("rmajor", scan=scan)
+    rminor = mfile_data.get("rminor", scan=scan)
+
+    # Convert normalized radius to actual radius
+    # rho ranges from -1 to 1, which corresponds to r = rmajor - rminor to rmajor + rminor
+    rho_ticks = np.array([-1, -0.75, -0.5, -0.25, 0, 0.25, 0.5, 0.75, 1])
+    r_ticks = rmajor + rho_ticks * rminor
+
+    ax2.set_xticks(rho_ticks)
+    ax2.set_xticklabels([f"{r:.2f}" for r in r_ticks])
+    ax2.set_xlabel("Radius [m]")
+    ax2.minorticks_on()
+    ax2.set_xlim(axis.get_xlim())
+
+    # Move secondary axis to the bottom
+    ax2.xaxis.set_ticks_position("bottom")
+    ax2.xaxis.set_label_position("bottom")
+    ax2.spines["bottom"].set_position(("outward", 30))
+
+    axis.legend()
+
+
+def plot_ion_frequency_profile(axis: plt.Axes, mfile_data: MFile, scan: int) -> None:
+    """Plot the ion thermal frequency profile on the given axis.
+
+    Parameters
+    ----------
+    axis : plt.Axes
+        Axis object to plot on
+    mfile_data : MFile
+        MFILE data object
+    scan : int
+        Scan number to use
+
+    """
+
+    freq_plasma_larmor_toroidal_deuteron_profile = [
+        mfile_data.data[f"freq_plasma_larmor_toroidal_deuteron_profile{i}"].get_scan(
+            scan
+        )
+        for i in range(
+            2 * int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan))
+        )
+    ]
+
+    freq_plasma_larmor_toroidal_triton_profile = [
+        mfile_data.data[f"freq_plasma_larmor_toroidal_triton_profile{i}"].get_scan(scan)
+        for i in range(
+            2 * int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan))
+        )
+    ]
+
+    axis.plot(
+        np.linspace(-1, 1, len(freq_plasma_larmor_toroidal_deuteron_profile)),
+        np.array(freq_plasma_larmor_toroidal_deuteron_profile) / 1e6,
+        color="red",
+        linestyle="-",
+        label=r"$f_{Larmor,toroidal,D}$",
+    )
+    axis.plot(
+        np.linspace(-1, 1, len(freq_plasma_larmor_toroidal_triton_profile)),
+        np.array(freq_plasma_larmor_toroidal_triton_profile) / 1e6,
+        color="green",
+        linestyle="-",
+        label=r"$f_{Larmor,toroidal,T}$",
+    )
+
+    axis.set_ylabel("Frequency [MHz]")
+    axis.set_xlabel("$\\rho \\ [r/a]$")
+    axis.grid(True, which="both", linestyle="--", alpha=0.5)
+    axis.minorticks_on()
+    axis.legend()
+
+
+def plot_plasma_coloumb_logarithms(axis: plt.Axes, mfile_data: MFile, scan: int) -> None:
+    """Plot the plasma coloumb logarithms on the given axis.
+
+    Parameters
+    ----------
+    axis : plt.Axes
+        Axis object to plot on
+    mfile_data : MFile
+        MFILE data object
+    scan : int
+        Scan number to use
+
+    """
+    plasma_coulomb_log_electron_electron_profile = [
+        mfile_data.data[f"plasma_coulomb_log_electron_electron_profile{i}"].get_scan(
+            scan
+        )
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    plasma_coulomb_log_electron_deuteron_profile = [
+        mfile_data.data[f"plasma_coulomb_log_electron_deuteron_profile{i}"].get_scan(
+            scan
+        )
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    plasma_coulomb_log_electron_triton_profile = [
+        mfile_data.data[f"plasma_coulomb_log_electron_triton_profile{i}"].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    plasma_coulomb_log_deuteron_triton_profile = [
+        mfile_data.data[f"plasma_coulomb_log_deuteron_triton_profile{i}"].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    plasma_coulomb_log_electron_alpha_thermal_profile = [
+        mfile_data.data[
+            f"plasma_coulomb_log_electron_alpha_thermal_profile{i}"
+        ].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    axis.plot(
+        np.linspace(0, 1, len(plasma_coulomb_log_electron_electron_profile)),
+        plasma_coulomb_log_electron_electron_profile,
+        color="blue",
+        linestyle="-",
+        label=r"$ln \Lambda_{e-e}$",
+    )
+
+    axis.plot(
+        np.linspace(0, 1, len(plasma_coulomb_log_electron_deuteron_profile)),
+        plasma_coulomb_log_electron_deuteron_profile,
+        color="pink",
+        linestyle="-",
+        label=r"$ln \Lambda_{e-D}$",
+    )
+
+    axis.plot(
+        np.linspace(0, 1, len(plasma_coulomb_log_electron_triton_profile)),
+        plasma_coulomb_log_electron_triton_profile,
+        color="green",
+        linestyle="-",
+        label=r"$ln \Lambda_{e-T}$",
+    )
+
+    axis.plot(
+        np.linspace(0, 1, len(plasma_coulomb_log_deuteron_triton_profile)),
+        plasma_coulomb_log_deuteron_triton_profile,
+        color="orange",
+        linestyle="-",
+        label=r"$ln \Lambda_{D-T}$",
+    )
+
+    axis.plot(
+        np.linspace(0, 1, len(plasma_coulomb_log_electron_alpha_thermal_profile)),
+        plasma_coulomb_log_electron_alpha_thermal_profile,
+        color="red",
+        linestyle="-",
+        label=r"$ln \Lambda_{e-\alpha,thermal}$",
+    )
+
+    axis.set_ylabel("Coulomb Logarithm")
+    axis.set_xlabel("$\\rho \\ [r/a]$")
+    axis.grid(True, which="both", linestyle="--", alpha=0.5)
+    axis.minorticks_on()
+    axis.legend()
+
+
+def plot_collision_time_profile(axis: plt.Axes, mfile_data: MFile, scan: int) -> None:
+    """Plot the plasma collision times on the given axis.
+
+    Parameters
+    ----------
+    axis : plt.Axes
+        Axis object to plot on
+    mfile_data : MFile
+        MFILE data object
+    scan : int
+        Scan number to use
+
+    """
+    t_plasma_electron_electron_collision_profile = [
+        mfile_data.data[f"t_plasma_electron_electron_collision_profile{i}"].get_scan(
+            scan
+        )
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    t_plasma_electron_deuteron_collision_profile = [
+        mfile_data.data[f"t_plasma_electron_deuteron_collision_profile{i}"].get_scan(
+            scan
+        )
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    t_plasma_electron_triton_collision_profile = [
+        mfile_data.data[f"t_plasma_electron_triton_collision_profile{i}"].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    t_plasma_electron_alpha_thermal_collision_profile = [
+        mfile_data.data[
+            f"t_plasma_electron_alpha_thermal_collision_profile{i}"
+        ].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    axis.plot(
+        np.linspace(0, 1, len(t_plasma_electron_electron_collision_profile)),
+        t_plasma_electron_electron_collision_profile,
+        color="blue",
+        linestyle="-",
+        label=r"$\tau_{e-e}$",
+    )
+
+    axis.plot(
+        np.linspace(0, 1, len(t_plasma_electron_deuteron_collision_profile)),
+        t_plasma_electron_deuteron_collision_profile,
+        color="pink",
+        linestyle="-",
+        label=r"$\tau_{e-D}$",
+    )
+
+    axis.plot(
+        np.linspace(0, 1, len(t_plasma_electron_triton_collision_profile)),
+        t_plasma_electron_triton_collision_profile,
+        color="green",
+        linestyle="-",
+        label=r"$\tau_{e-T}$",
+    )
+
+    axis.plot(
+        np.linspace(0, 1, len(t_plasma_electron_alpha_thermal_collision_profile)),
+        t_plasma_electron_alpha_thermal_collision_profile,
+        color="red",
+        linestyle="-",
+        label=r"$\tau_{e-\alpha,thermal}$",
+    )
+
+    axis.set_yscale("log")
+    axis.set_ylabel("Collision Time [s]")
+    axis.set_xlabel("$\\rho \\ [r/a]$")
+    axis.grid(True, which="both", linestyle="--", alpha=0.5)
+    axis.minorticks_on()
+    axis.legend()
+
+
+def plot_collision_frequency_profile(
+    axis: plt.Axes, mfile_data: MFile, scan: int
+) -> None:
+    """Plot the plasma collision frequencies on the given axis.
+
+    Parameters
+    ----------
+    axis : plt.Axes
+        Axis object to plot on
+    mfile_data : MFile
+        MFILE data object
+    scan : int
+        Scan number to use
+
+    """
+    freq_plasma_electron_electron_collision_profile = [
+        mfile_data.data[f"freq_plasma_electron_electron_collision_profile{i}"].get_scan(
+            scan
+        )
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    freq_plasma_electron_deuteron_collision_profile = [
+        mfile_data.data[f"freq_plasma_electron_deuteron_collision_profile{i}"].get_scan(
+            scan
+        )
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    freq_plasma_electron_triton_collision_profile = [
+        mfile_data.data[f"freq_plasma_electron_triton_collision_profile{i}"].get_scan(
+            scan
+        )
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    freq_plasma_electron_alpha_thermal_collision_profile = [
+        mfile_data.data[
+            f"freq_plasma_electron_alpha_thermal_collision_profile{i}"
+        ].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    axis.plot(
+        np.linspace(0, 1, len(freq_plasma_electron_electron_collision_profile)),
+        freq_plasma_electron_electron_collision_profile,
+        color="blue",
+        linestyle="-",
+        label=r"$\nu_{e-e}$",
+    )
+
+    axis.plot(
+        np.linspace(0, 1, len(freq_plasma_electron_deuteron_collision_profile)),
+        freq_plasma_electron_deuteron_collision_profile,
+        color="pink",
+        linestyle="-",
+        label=r"$\nu_{e-D}$",
+    )
+
+    axis.plot(
+        np.linspace(0, 1, len(freq_plasma_electron_triton_collision_profile)),
+        freq_plasma_electron_triton_collision_profile,
+        color="green",
+        linestyle="-",
+        label=r"$\nu_{e-T}$",
+    )
+
+    axis.plot(
+        np.linspace(0, 1, len(freq_plasma_electron_alpha_thermal_collision_profile)),
+        freq_plasma_electron_alpha_thermal_collision_profile,
+        color="red",
+        linestyle="-",
+        label=r"$\nu_{e-\alpha,thermal}$",
+    )
+    axis.set_yscale("log")
+    axis.set_ylabel("Collision Frequency [Hz]")
+    axis.set_xlabel("$\\rho \\ [r/a]$")
+    axis.grid(True, which="both", linestyle="--", alpha=0.5)
+    axis.minorticks_on()
+    axis.legend()
+
+
+def plot_mean_free_path_profile(axis: plt.Axes, mfile_data: MFile, scan: int) -> None:
+    """Plot the plasma mean free path on the given axis.
+
+    Parameters
+    ----------
+    axis : plt.Axes
+        Axis object to plot on
+    mfile_data : MFile
+        MFILE data object
+    scan : int
+        Scan number to use
+
+    """
+    len_plasma_electron_electron_mean_free_path_profile = [
+        mfile_data.data[
+            f"len_plasma_electron_electron_mean_free_path_profile{i}"
+        ].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    len_plasma_electron_deuteron_mean_free_path_profile = [
+        mfile_data.data[
+            f"len_plasma_electron_deuteron_mean_free_path_profile{i}"
+        ].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    len_plasma_electron_triton_mean_free_path_profile = [
+        mfile_data.data[
+            f"len_plasma_electron_triton_mean_free_path_profile{i}"
+        ].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    len_plasma_electron_alpha_thermal_mean_free_path_profile = [
+        mfile_data.data[
+            f"len_plasma_electron_alpha_thermal_mean_free_path_profile{i}"
+        ].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    axis.plot(
+        np.linspace(0, 1, len(len_plasma_electron_electron_mean_free_path_profile)),
+        len_plasma_electron_electron_mean_free_path_profile,
+        color="blue",
+        linestyle="-",
+        label=r"$\lambda_{mfp,e-e}$",
+    )
+
+    axis.plot(
+        np.linspace(0, 1, len(len_plasma_electron_deuteron_mean_free_path_profile)),
+        len_plasma_electron_deuteron_mean_free_path_profile,
+        color="pink",
+        linestyle="-",
+        label=r"$\lambda_{mfp,e-D}$",
+    )
+
+    axis.plot(
+        np.linspace(0, 1, len(len_plasma_electron_triton_mean_free_path_profile)),
+        len_plasma_electron_triton_mean_free_path_profile,
+        color="green",
+        linestyle="-",
+        label=r"$\lambda_{mfp,e-T}$",
+    )
+    axis.plot(
+        np.linspace(0, 1, len(len_plasma_electron_alpha_thermal_mean_free_path_profile)),
+        len_plasma_electron_alpha_thermal_mean_free_path_profile,
+        color="red",
+        linestyle="-",
+        label=r"$\lambda_{mfp,e-\alpha,thermal}$",
+    )
+    axis.set_yscale("log")
+    axis.set_ylabel("Mean Free Path [m]")
+    axis.set_xlabel("$\\rho \\ [r/a]$")
+    axis.grid(True, which="both", linestyle="--", alpha=0.5)
+    axis.minorticks_on()
+    axis.legend()
+
+
+def plot_ion_slowing_down_time_profile(
+    axis: plt.Axes, mfile_data: MFile, scan: int
+) -> None:
+    """Plot the plasma Spitzer slowing down time on the given axis.
+
+    Parameters
+    ----------
+    axis : plt.Axes
+        Axis object to plot on
+    mfile_data : MFile
+        MFILE data object
+    scan : int
+        Scan number to use
+
+    """
+    t_plasma_electron_alpha_spitzer_slow_profile = [
+        mfile_data.data[f"t_plasma_electron_alpha_spitzer_slow_profile{i}"].get_scan(
+            scan
+        )
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    axis.plot(
+        np.linspace(0, 1, len(t_plasma_electron_alpha_spitzer_slow_profile)),
+        t_plasma_electron_alpha_spitzer_slow_profile,
+        color="red",
+        linestyle="-",
+        label=r"$\tau_{e-\alpha,Spitzer}$",
+    )
+
+    axis.set_yscale("log")
+    axis.set_ylabel("Spitzer Slowing Down Time [s]")
+    axis.set_xlabel("$\\rho \\ [r/a]$")
+    axis.grid(True, which="both", linestyle="--", alpha=0.5)
+    axis.minorticks_on()
+    axis.legend()
+
+
+def plot_resistivity_profile(axis: plt.Axes, mfile_data: MFile, scan: int) -> None:
+    """Plot the plasma resistivity on the given axis.
+
+    Parameters
+    ----------
+    axis : plt.Axes
+        Axis object to plot on
+    mfile_data : MFile
+        MFILE data object
+    scan : int
+        Scan number to use
+
+    """
+    res_plasma_spitzer_profile = [
+        mfile_data.data[f"res_plasma_spitzer_profile{i}"].get_scan(scan)
+        for i in range(int(mfile_data.data["n_plasma_profile_elements"].get_scan(scan)))
+    ]
+
+    axis.plot(
+        np.linspace(0, 1, len(res_plasma_spitzer_profile)),
+        res_plasma_spitzer_profile,
+        color="red",
+        linestyle="-",
+        label=r"$\eta_{Spitzer}$",
+    )
+
+    axis.set_yscale("log")
+    axis.set_ylabel("Resistivity [Ohm m]")
+    axis.set_xlabel("$\\rho \\ [r/a]$")
+    axis.grid(True, which="both", linestyle="--", alpha=0.5)
+    axis.minorticks_on()
+    axis.legend()
+
+
 def plot_equality_constraint_equations(axis: plt.Axes, m_file_data: MFile, scan: int):
     """Plot the equality constraints for a solution and their normalised residuals
 
@@ -13391,31 +14121,25 @@ def main_plot(
     plot_density_limit_comparison(figs[15].add_subplot(221), m_file, scan)
     plot_confinement_time_comparison(figs[16].add_subplot(224), m_file, scan)
 
-    DetailedPhysics.plot_debye_length_profile(figs[17].add_subplot(232), m_file, scan)
-    DetailedPhysics.plot_velocity_profile(figs[17].add_subplot(233), m_file, scan)
-    DetailedPhysics.plot_plasma_coloumb_logarithms(
-        figs[17].add_subplot(231), m_file, scan
-    )
-    DetailedPhysics.plot_collision_time_profile(figs[17].add_subplot(234), m_file, scan)
-    DetailedPhysics.plot_collision_frequency_profile(
-        figs[17].add_subplot(235), m_file, scan
-    )
-    DetailedPhysics.plot_mean_free_path_profile(figs[17].add_subplot(236), m_file, scan)
+    plot_debye_length_profile(figs[17].add_subplot(232), m_file, scan)
+    plot_velocity_profile(figs[17].add_subplot(233), m_file, scan)
+    plot_plasma_coloumb_logarithms(figs[17].add_subplot(231), m_file, scan)
+    plot_collision_time_profile(figs[17].add_subplot(234), m_file, scan)
+    plot_collision_frequency_profile(figs[17].add_subplot(235), m_file, scan)
+    plot_mean_free_path_profile(figs[17].add_subplot(236), m_file, scan)
 
-    DetailedPhysics.plot_ion_slowing_down_time_profile(
-        figs[18].add_subplot(231), m_file, scan
-    )
+    plot_ion_slowing_down_time_profile(figs[18].add_subplot(231), m_file, scan)
 
-    DetailedPhysics.plot_resistivity_profile(figs[18].add_subplot(232), m_file, scan)
+    plot_resistivity_profile(figs[18].add_subplot(232), m_file, scan)
 
     ax_electron_freq = figs[19].add_subplot(211)
-    DetailedPhysics.plot_electron_frequency_profile(ax_electron_freq, m_file, scan)
+    plot_electron_frequency_profile(ax_electron_freq, m_file, scan)
 
     ax_ion_freq = figs[19].add_subplot(413, sharex=ax_electron_freq)
-    DetailedPhysics.plot_ion_frequency_profile(ax_ion_freq, m_file, scan)
+    plot_ion_frequency_profile(ax_ion_freq, m_file, scan)
 
     ax_larmor = figs[19].add_subplot(414, sharex=ax_electron_freq)
-    DetailedPhysics.plot_larmor_radius_profile(ax_larmor, m_file, scan)
+    plot_larmor_radius_profile(ax_larmor, m_file, scan)
 
     figs[19].subplots_adjust(hspace=0.5)
